@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net472;net5.0</TargetFrameworks>
+    <TargetFrameworks>net472;net6.0</TargetFrameworks>
     <Version>0.0.0</Version>
   </PropertyGroup>
 
@@ -10,10 +10,10 @@
   </PropertyGroup>
 
   <!-- .NET Standard 2.0 references, compilation flags and build options -->
-  <PropertyGroup Condition=" '$(TargetFramework)' == 'net5.0'">
+  <PropertyGroup Condition=" '$(TargetFramework)' == 'net6.0'">
     <DefineConstants>NODATIME2;AUTOFIXTURE4</DefineConstants>
   </PropertyGroup>
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net5.0'">
+  <ItemGroup Condition=" '$(TargetFramework)' == 'net6.0'">
 	  <PackageReference Include="AutoFixture" Version="4.0.0-rc1" />
 	  <PackageReference Include="FluentAssertions" Version="4.19.4" />
 	  <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.5.0" />
